refactor: change folder name to SupplierDashboard
